Crucial changes around value and entity display name
previous implementation was not good enough. It was a var instance property, which caused some confusion. The fact that it is an instance ‘var’ means it’s possible to change display name during life time of the instance, while it’s supposed to be a fixed value in real apps. Also, it meant different instances of the same type could easily have different display name, which is not a desired behaviour. So the ‘displayName’ property really should be an attribute of the type, not instance. For example, a ‘User’ entity should have intrinsic display name ‘User’ but in context of a ‘Comment’ it may want to be displayed as “Author”, but it should be consistent across all comments.