Important last-minute changes in validation implementation details
the ‘validValue’ helpers are now available on ‘Optional’ level if the ‘wrapped’ type is a ‘ValueWrapper’ & ‘Mandatory’, and on ‘ValueWrapper’ level only if it’s explicitly marked as ‘NonMandatory’. So when we put a wrapper inside ‘Optional’ - we don’t have same ‘validValue’ functions producing different results and have much better developer experience. For ‘NonRequired*<*>?’ types we say ‘try optionalWrapper?.validValue()’ to access valid value or throw, because if it’s empty - we consider it as valid anyway and should return ‘nil’ without throwing an error - this is exactly what the provided API will do, plus we leverage default usage and autocompletion ofr ‘Optional’. For ‘Required*<*>?’ types we say ‘try optionalWrapper.validValue()’ to access valid value or throw, if it’s invalid or EMPTY, because teh wrapper marked as ‘Mandatory’ and it only makes sense to access valid value from ‘Optional’ level - so we properly react on both empty value or invalid non-empty, or return NON-optional value (that’s important for convenience, sicne we declared the type as ‘Mandatory’); also ‘validValue(_ collectErrors:…)’ returns implicitly unwrapped optional to keep API consistent. Plus many last minute fixes and improvements for ‘Optional’ support.